refactor : King이 잡히면 게임 종료하는 기능 수정
- 컨트롤러에서 게임을 종료시키도록 수정
