+++ src/main/java/domain/piece/Rook.java
@@ -39,18 +39,6 @@ public class Rook extends Piece {
 	}
 
 	@Override
-	public void move(Position targetPosition, List<Rank> ranks) {
-		Optional<Piece> piece = hasPieceInBoard(ranks, targetPosition);
-		piece.ifPresent(targetPiece -> {
-			if (targetPiece.team.equals(this.team)) {
-				throw new InvalidPositionException(InvalidPositionException.HAS_OUR_TEAM_AT_TARGET_POSITION);
-			}
-			capture(targetPiece, ranks);
-		});
-		this.changePosition(targetPosition, ranks);
-	}
-
-	@Override
 	protected String getSymbol() {
 		return SYMBOL;
 	}
