+++ src/main/java/domain/piece/team/Team.java
@@ -1,7 +1,7 @@
 package domain.piece.team;
 
 public enum Team {
-	BLACK(), WHITE();
+	BLACK, WHITE;
 
 	public static Team changeTurn(Team turn) {
 		if (WHITE == turn) {
